"Refactor and optimize bot's services and clients"
Remove TinkoffExchanger due to its redundancy. Rename Binance P2P assets and correct spelling errors. Update the CultureInfo naming convention in MirPayRate.cs and added in null-checks for inputs. Correct the selector paths to be sure they are targeting correctly. Changes were also made to the decimal rounding methodology in RatesService.cs for more accurate results. Sealed the PochtaBankRate class for performance improvements and input validation was added. Enhancements made to the BinanceP2PExchanger.cs class include adding name and asset as constructor parameters, introducing the option for random pauses during the process, and adjusting parameters for the Binance Client methods. Add a null-check for the PochtaBank in the QiwiExchanger's constructor and replace the CBR client dependency with the PochtaBankRate. Include new services in the HostBuilder configuration in Program.cs. Lastly, added new Freedom Finance's client and updated the project file dependencies.